+++ pkg/accounts/accounts_test.go
@@ -1,4 +1,4 @@
-package internal
+package accounts
 
 import (
 	"testing"
