+++ pkg/accounts/simpleAccountant.go
@@ -1,4 +1,4 @@
-package internal
+package accounts
 
 import (
 	"fmt"
